Clean up source tree, Makefile, and fix complier warnings
The main Makefile has been cleaned up quite a bit:

  - separate Makefiles for creating native packages and tarballs.
  - additional CFLAGS have been added to fix complier warnings.
  - version is now hard-coded. No longer using `git describe`.
  - files included in the tarball package are now hard-coded. No
    longer using `git ls-files`.
  - additional clean targets.

All source files have been relocated under the src directory.

Several complier warnings caused by using deprecated LDAP functions
have been fixed. The source tree compiles without warnings on both
Debian and RHEL based systems.

+++ src/utils.h
@@ -12,7 +12,7 @@ struct pkfs_config {
   const char *uri;
   const char *pass;
   const char *key_attr;
-  long timeout;
+  long long timeout;
 };
 
 typedef struct pkfs_config pkfs_config_t;
@@ -21,3 +21,4 @@ void initialize_config(void);
 void uid_from_path(const char *path, char **uid);
 
 #endif
+
